backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
librosa
matplotlib
numpy
scipy
fastapi-cors
dotenv-python